save works, working on including variable into save string
